predictor/requirements.txt
affine==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.17
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.0
cycler==0.12.1
deepforest==1.3.3
dnspython==2.6.1
email-validator==2.2.0
eval-type-backport==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
imageio==2.35.1
jinja2==3.1.4
kiwisolver==1.4.7
lazy-loader==0.4
lightning-utilities==0.11.7
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pika==1.3.2
pillow==10.4.0
psutil==6.0.0
pycocotools==2.0.8
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2
pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytorch-lightning==2.4.0
pytz==2024.2
pyyaml==6.0.2
rasterio==1.4.0
rich==13.8.1
rtree==1.3.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
slidingwindow==0.0.14
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tifffile==2024.9.20
torch==2.3.1+cpu
torchmetrics==1.0.3
torchvision==0.18.1+cpu
tqdm==4.66.5
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
xmltodict==0.13.0
yarl==1.13.1
